requirements.txt
APScheduler==3.9.1
certifi==2022.5.18.1
charset-normalizer==2.0.12
click==8.1.3
fasttext==0.9.2
Flask==2.1.2
gensim==4.2.0
greenlet==1.1.2
idna==3.3
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
numpy==1.22.4
psycopg2-binary==2.9.3
pybind11==2.9.2
python-dotenv==0.20.0
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
requests==2.28.0
scipy==1.8.1
six==1.16.0
smart-open==6.0.0
SQLAlchemy==1.4.37
tqdm==4.64.0
tzdata==2022.1
tzlocal==4.2
urllib3==1.26.9
Werkzeug==2.1.2
